z3: collapse per-sort to_z3/from_z3 into single implementations
Per PR review: merge the four astext/{bool,bv,float,string}.rs modules into one
to_z3 (a single match over the unified op enum; polymorphic Not/And/Or/Xor pick
the bool vs bitvector Z3 constructor from the node type) and one from_z3 (a
single match over the Z3 decl kind, with Ite/uninterpreted-constants/numerals
choosing the result sort from the Z3 sort kind). Removes the separate files;
shared helpers (fprm/fsort conversion, bv2int, unicode decode) move into
astext.rs. All 334 z3 round-trip tests still pass.
