--- crates/clarirs_z3/src/astext/bv.rs
@@ -1,366 +1,0 @@
-use clarirs_core::prelude::*;
-use clarirs_z3_sys as z3;
-
-use super::AstExtZ3;
-use crate::{Z3_CONTEXT, astext::child, check_z3_error, rc::RcAst};
-
-pub(crate) fn to_z3(ast: &AstRef, children: &[RcAst]) -> Result<RcAst, ClarirsError> {
-    Z3_CONTEXT.with(|&z3_ctx| unsafe {
-        Ok(match ast.op() {
-            AstOp::BVS(s, w) => {
-                let s_cstr = std::ffi::CString::new(s.as_str()).unwrap();
-                let sym = z3::mk_string_symbol(z3_ctx, s_cstr.as_ptr());
-                let sort = z3::mk_bv_sort(z3_ctx, *w);
-                RcAst::try_from(z3::mk_const(z3_ctx, sym, sort))?
-            }
-            AstOp::BVV(v) => {
-                let sort = z3::mk_bv_sort(z3_ctx, v.len());
-                let numeral = v.to_biguint().to_string();
-                let numeral_cstr = std::ffi::CString::new(numeral).unwrap();
-                RcAst::try_from(z3::mk_numeral(z3_ctx, numeral_cstr.as_ptr(), sort))?
-            }
-            AstOp::Not(..) => unop!(z3_ctx, children, mk_bvnot),
-            AstOp::Neg(..) => unop!(z3_ctx, children, mk_bvneg),
-            AstOp::And(..) => naryop!(z3_ctx, children, mk_bvand),
-            AstOp::Or(..) => naryop!(z3_ctx, children, mk_bvor),
-            AstOp::Xor(..) => naryop!(z3_ctx, children, mk_bvxor),
-            AstOp::Add(..) => naryop!(z3_ctx, children, mk_bvadd),
-            AstOp::Sub(..) => binop!(z3_ctx, children, mk_bvsub),
-            AstOp::Mul(..) => naryop!(z3_ctx, children, mk_bvmul),
-            AstOp::UDiv(..) => binop!(z3_ctx, children, mk_bvudiv),
-            AstOp::SDiv(..) => binop!(z3_ctx, children, mk_bvsdiv),
-            AstOp::URem(..) => binop!(z3_ctx, children, mk_bvurem),
-            AstOp::SRem(..) => binop!(z3_ctx, children, mk_bvsrem),
-            AstOp::ShL(..) => binop!(z3_ctx, children, mk_bvshl),
-            AstOp::LShR(..) => binop!(z3_ctx, children, mk_bvlshr),
-            AstOp::AShR(..) => binop!(z3_ctx, children, mk_bvashr),
-            AstOp::RotateLeft(..) => binop!(z3_ctx, children, mk_ext_rotate_left),
-            AstOp::RotateRight(..) => binop!(z3_ctx, children, mk_ext_rotate_right),
-            AstOp::ZeroExt(_, i) => {
-                RcAst::try_from(z3::mk_zero_ext(z3_ctx, *i, **child(children, 0)?))?
-            }
-            AstOp::SignExt(_, i) => {
-                RcAst::try_from(z3::mk_sign_ext(z3_ctx, *i, **child(children, 0)?))?
-            }
-            AstOp::Extract(a, high, low) => {
-                if high >= &a.size() {
-                    return Err(ClarirsError::ConversionError(
-                        "high index is greater than bitvector size".to_string(),
-                    ));
-                }
-                if low >= &a.size() {
-                    return Err(ClarirsError::ConversionError(
-                        "low index is greater than bitvector size".to_string(),
-                    ));
-                }
-                if low > high {
-                    return Err(ClarirsError::ConversionError(
-                        "low index is greater than high index".to_string(),
-                    ));
-                }
-                let a = child(children, 0)?;
-                RcAst::try_from(z3::mk_extract(z3_ctx, *high, *low, **a))?
-            }
-            AstOp::Concat(args) => {
-                // Z3's concat constructor is binary, so we chain them
-                if args.is_empty() {
-                    return Err(ClarirsError::InvalidArguments(
-                        "Concat requires at least one argument".to_string(),
-                    ));
-                }
-                let mut result = child(children, 0)?.clone();
-                for i in 1..children.len() {
-                    result =
-                        RcAst::try_from(z3::mk_concat(z3_ctx, *result, **child(children, i)?))?;
-                }
-                result
-            }
-            AstOp::ByteReverse(a) => {
-                let size = a.size();
-                if size == 0 || size % 8 != 0 {
-                    return Err(ClarirsError::ConversionError(
-                        "reverse only supports bitvectors with size multiple of 8".to_string(),
-                    ));
-                }
-
-                let child_z3 = child(children, 0)?;
-                let num_bytes = size / 8;
-
-                // Extract the last byte (lowest bits) as the initial accumulator
-                let mut result = RcAst::try_from(z3::mk_extract(z3_ctx, 7, 0, **child_z3))?;
-
-                // Extract remaining bytes in reverse order and concat
-                for i in 1..num_bytes {
-                    let high = (i + 1) * 8 - 1;
-                    let low = i * 8;
-                    let byte = RcAst::try_from(z3::mk_extract(z3_ctx, high, low, **child_z3))?;
-                    result = RcAst::try_from(z3::mk_concat(z3_ctx, *result, *byte))?;
-                }
-
-                result
-            }
-            AstOp::ITE(..) => {
-                let cond = child(children, 0)?;
-                let then = child(children, 1)?;
-                let else_ = child(children, 2)?;
-                RcAst::try_from(z3::mk_ite(z3_ctx, **cond, **then, **else_))?
-            }
-            AstOp::FpToIEEEBV(..) => {
-                let a = child(children, 0)?;
-                RcAst::try_from(z3::mk_fpa_to_ieee_bv(z3_ctx, **a))?
-            }
-            AstOp::FpToUBV(_, size, rm) => {
-                let rm_ast = super::float::fprm_to_z3(*rm)?;
-                let a = child(children, 0)?;
-                RcAst::try_from(z3::mk_fpa_to_ubv(z3_ctx, *rm_ast, **a, *size))?
-            }
-            AstOp::FpToSBV(_, size, rm) => {
-                let rm_ast = super::float::fprm_to_z3(*rm)?;
-                let a = child(children, 0)?;
-                RcAst::try_from(z3::mk_fpa_to_sbv(z3_ctx, *rm_ast, **a, *size))?
-            }
-            AstOp::StrLen(..) => {
-                let a = child(children, 0)?;
-                let str_len = RcAst::try_from(z3::mk_seq_length(z3_ctx, **a))?;
-                RcAst::try_from(z3::mk_int2bv(z3_ctx, 64, *str_len))?
-            }
-            AstOp::StrIndexOf(..) => {
-                let haystack = child(children, 0)?;
-                let needle = child(children, 1)?;
-                let offset_bv = child(children, 2)?;
-                let offset_int = RcAst::try_from(z3::mk_bv2int(z3_ctx, **offset_bv, false))?;
-                let index_int =
-                    RcAst::try_from(z3::mk_seq_index(z3_ctx, **haystack, **needle, *offset_int))?;
-                RcAst::try_from(z3::mk_int2bv(z3_ctx, 64, *index_int))?
-            }
-            AstOp::StrToBV(..) => {
-                let a = child(children, 0)?;
-                let int_val = RcAst::try_from(z3::mk_str_to_int(z3_ctx, **a))?;
-                RcAst::try_from(z3::mk_int2bv(z3_ctx, 64, *int_val))?
-            }
-            AstOp::Union(..) | AstOp::Intersection(..) | AstOp::Widen(..) => {
-                // These are not supported in Z3
-                return Err(ClarirsError::ConversionError(
-                    "vsa types are not currently supported in the z3 backend".to_string(),
-                ));
-            }
-            _ => unreachable!("non-bitvector op in bv::to_z3"),
-        })
-        .and_then(|maybe_null| {
-            check_z3_error()?;
-            Ok(maybe_null)
-        })
-    })
-}
-
-pub(crate) fn from_z3<'c>(
-    ctx: &'c Context<'c>,
-    ast: impl Into<RcAst>,
-) -> Result<AstRef<'c>, ClarirsError> {
-    Z3_CONTEXT.with(|z3_ctx| unsafe {
-        let ast = ast.into();
-        let ast_kind = z3::get_ast_kind(*z3_ctx, *ast);
-        match ast_kind {
-            z3::AstKind::Numeral => {
-                let numeral_string = z3::get_numeral_string(*z3_ctx, *ast);
-                let numeral_str = std::ffi::CStr::from_ptr(numeral_string).to_str().unwrap();
-                let sort = z3::get_sort(*z3_ctx, *ast);
-                let sort_num = z3::get_bv_sort_size(*z3_ctx, sort);
-                let biguint = BitVec::from_str(numeral_str, sort_num).unwrap();
-                ctx.bvv(biguint)
-            }
-            z3::AstKind::App => {
-                let app = z3::to_app(*z3_ctx, *ast);
-                let decl = z3::get_app_decl(*z3_ctx, app);
-                let decl_kind = z3::get_decl_kind(*z3_ctx, decl);
-                let sort = z3::get_sort(*z3_ctx, *ast);
-                let width = z3::get_bv_sort_size(*z3_ctx, sort);
-
-                match decl_kind {
-                    z3::DeclKind::Uninterpreted => {
-                        let sym = z3::get_decl_name(*z3_ctx, decl);
-                        let name = z3::get_symbol_string(*z3_ctx, sym);
-                        let name = std::ffi::CStr::from_ptr(name).to_str().unwrap();
-                        ctx.bvs(name, width as u32)
-                    }
-                    z3::DeclKind::Bnot => {
-                        let arg = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let inner = AstRef::from_z3(ctx, arg)?;
-                        ctx.not(inner)
-                    }
-                    z3::DeclKind::Bneg => {
-                        let arg = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let inner = AstRef::from_z3(ctx, arg)?;
-                        ctx.neg(inner)
-                    }
-                    z3::DeclKind::Band
-                    | z3::DeclKind::Bor
-                    | z3::DeclKind::Bxor
-                    | z3::DeclKind::Badd
-                    | z3::DeclKind::Bmul => {
-                        // N-ary ops: collect all args into a Vec
-                        let num_args = z3::get_app_num_args(*z3_ctx, app);
-                        let mut args = Vec::with_capacity(num_args as usize);
-                        for i in 0..num_args {
-                            let arg = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, i))?;
-                            args.push(AstRef::from_z3(ctx, arg)?);
-                        }
-                        match decl_kind {
-                            z3::DeclKind::Band => ctx.and(args),
-                            z3::DeclKind::Bor => ctx.or(args),
-                            z3::DeclKind::Bxor => ctx.xor(args),
-                            z3::DeclKind::Badd => ctx.add_many(args),
-                            z3::DeclKind::Bmul => ctx.mul_many(args),
-                            _ => unreachable!(),
-                        }
-                    }
-                    z3::DeclKind::Bsub
-                    | z3::DeclKind::Budiv
-                    | z3::DeclKind::Bsdiv
-                    | z3::DeclKind::Burem
-                    | z3::DeclKind::Bsrem
-                    | z3::DeclKind::Bshl
-                    | z3::DeclKind::Blshr
-                    | z3::DeclKind::Bashr => {
-                        let arg0 = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let arg1 = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 1))?;
-                        let a = AstRef::from_z3(ctx, arg0)?;
-                        let b = AstRef::from_z3(ctx, arg1)?;
-                        match decl_kind {
-                            z3::DeclKind::Bsub => ctx.sub(a, b),
-                            z3::DeclKind::Budiv => ctx.udiv(a, b),
-                            z3::DeclKind::Bsdiv => ctx.sdiv(a, b),
-                            z3::DeclKind::Burem => ctx.urem(a, b),
-                            z3::DeclKind::Bsrem => ctx.srem(a, b),
-                            z3::DeclKind::Bshl => ctx.shl(a, b),
-                            z3::DeclKind::Blshr => ctx.lshr(a, b),
-                            z3::DeclKind::Bashr => ctx.ashr(a, b),
-                            _ => unreachable!(),
-                        }
-                    }
-                    z3::DeclKind::ExtRotateLeft | z3::DeclKind::ExtRotateRight => {
-                        let arg0 = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let arg1 = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 1))?;
-                        let a = AstRef::from_z3(ctx, arg0)?;
-                        let b = AstRef::from_z3(ctx, arg1)?;
-                        match decl_kind {
-                            z3::DeclKind::ExtRotateLeft => ctx.rotate_left(a, b),
-                            z3::DeclKind::ExtRotateRight => ctx.rotate_right(a, b),
-                            _ => unreachable!(),
-                        }
-                    }
-                    z3::DeclKind::ZeroExt | z3::DeclKind::SignExt => {
-                        let arg = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let inner = AstRef::from_z3(ctx, arg)?;
-                        let i = z3::get_decl_int_parameter(*z3_ctx, decl, 0) as u32;
-                        match decl_kind {
-                            z3::DeclKind::ZeroExt => ctx.zero_ext(&inner, i),
-                            z3::DeclKind::SignExt => ctx.sign_ext(inner, i),
-                            _ => unreachable!(),
-                        }
-                    }
-                    z3::DeclKind::Extract => {
-                        let arg = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let inner = AstRef::from_z3(ctx, arg)?;
-                        let high = z3::get_decl_int_parameter(*z3_ctx, decl, 0) as u32;
-                        let low = z3::get_decl_int_parameter(*z3_ctx, decl, 1) as u32;
-                        ctx.extract(inner, high, low)
-                    }
-                    z3::DeclKind::Concat => {
-                        let num_args = z3::get_app_num_args(*z3_ctx, app);
-                        let mut concat_args = Vec::with_capacity(num_args as usize);
-                        for i in 0..num_args {
-                            let arg = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, i))?;
-                            let val = AstRef::from_z3(ctx, arg)?;
-                            concat_args.push(val);
-                        }
-                        ctx.concat(concat_args)
-                    }
-                    z3::DeclKind::Ite => {
-                        let cond = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                        let then = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 1))?;
-                        let else_ = RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 2))?;
-                        let cond = AstRef::from_z3(ctx, cond)?;
-                        let then = AstRef::from_z3(ctx, then)?;
-                        let else_ = AstRef::from_z3(ctx, else_)?;
-                        ctx.ite(cond, then, else_)
-                    }
-                    // Special case for bv2int used in string operations
-                    z3::DeclKind::Int2bv => {
-                        let inner_int = z3::get_app_arg(*z3_ctx, app, 0);
-                        let inner_ast_kind = z3::get_ast_kind(*z3_ctx, *ast);
-                        match inner_ast_kind {
-                            z3::AstKind::Numeral => {
-                                let numeral_string = z3::get_numeral_string(*z3_ctx, inner_int);
-                                let numeral_str =
-                                    std::ffi::CStr::from_ptr(numeral_string).to_str().unwrap();
-                                let sort = z3::get_sort(*z3_ctx, inner_int);
-                                let sort_num = z3::get_bv_sort_size(*z3_ctx, sort);
-                                let biguint = BitVec::from_str(numeral_str, sort_num).unwrap();
-                                ctx.bvv(biguint)
-                            }
-                            z3::AstKind::App => {
-                                let app = z3::to_app(*z3_ctx, inner_int);
-                                let decl = z3::get_app_decl(*z3_ctx, app);
-                                let decl_kind = z3::get_decl_kind(*z3_ctx, decl);
-
-                                match decl_kind {
-                                    z3::DeclKind::Bv2int => {
-                                        let arg =
-                                            RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                                        AstRef::from_z3(ctx, arg)
-                                    }
-                                    z3::DeclKind::SeqIndex => {
-                                        // int2bv(seq.indexof(haystack, needle, offset))
-                                        let arg0 =
-                                            RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                                        let arg1 =
-                                            RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 1))?;
-                                        let arg2 =
-                                            RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 2))?;
-                                        let haystack = AstRef::from_z3(ctx, arg0)?;
-                                        let needle = AstRef::from_z3(ctx, arg1)?;
-                                        // offset is an int, convert to bv
-                                        let offset_bv =
-                                            RcAst::try_from(z3::mk_int2bv(*z3_ctx, 64, *arg2))?;
-                                        let offset_simplified =
-                                            RcAst::try_from(z3::simplify(*z3_ctx, *offset_bv))?;
-                                        let offset = AstRef::from_z3(ctx, offset_simplified)?;
-                                        ctx.str_index_of(haystack, needle, offset)
-                                    }
-                                    z3::DeclKind::StrToInt => {
-                                        // int2bv(str.to_int(s))
-                                        let arg0 =
-                                            RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                                        let s = AstRef::from_z3(ctx, arg0)?;
-                                        ctx.str_to_bv(s)
-                                    }
-                                    z3::DeclKind::SeqLength => {
-                                        // int2bv(seq.len(s))
-                                        let arg0 =
-                                            RcAst::try_from(z3::get_app_arg(*z3_ctx, app, 0))?;
-                                        let s = AstRef::from_z3(ctx, arg0)?;
-                                        ctx.str_len(s)
-                                    }
-                                    _ => Err(ClarirsError::ConversionError(format!(
-                                        "unexpected inner decl kind in Int2bv: {:?}",
-                                        decl_kind,
-                                    ))),
-                                }
-                            }
-                            _ => Err(ClarirsError::ConversionError(
-                                "expected a numeral or bv2int".to_string(),
-                            )),
-                        }
-                    }
-                    _ => Err(ClarirsError::ConversionError(
-                        "Failed converting from z3: unknown decl kind for bitvec".to_string(),
-                    )),
-                }
-            }
-            _ => Err(ClarirsError::ConversionError(
-                "Failed converting from z3: unknown ast kind for bitvec".to_string(),
-            )),
-        }
-    })
-}
